refactor serializers to resolve circular dependencies
