+++ api/v2/serializers/__init__.py
@@ -1,12 +1,27 @@
-from .allocation import AllocationSerializer
-from .identity import IdentitySerializer, IdentitySummarySerializer
-from .image import ImageSerializer, ImageSummarySerializer, ImageBookmarkSerializer
-from .instance import InstanceSerializer, InstanceSummarySerializer
-from .project import ProjectSerializer
-from .provider import ProviderSerializer, ProviderSummarySerializer, ProviderTypeSerializer, PlatformTypeSerializer
-from .provider_machine import ProviderMachineSerializer, ProviderMachineSummarySerializer
-from .quota import QuotaSerializer
-from .size import SizeSerializer, SizeSummarySerializer
-from .tag import TagSerializer
-from .user import UserSerializer
-from .volume import VolumeSerializer, VolumeSummarySerializer
+from .details import *
+from .summaries import *
+
+# from .details import AllocationSerializer
+# from .details import IdentitySerializer
+# from .details import ImageSerializer
+# from .details import ImageBookmarkSerializer
+# from .details import InstanceSerializer
+# from .details import ProjectSerializer
+# from .details import ProviderSerializer, ProviderTypeSerializer, PlatformTypeSerializer
+# from .details import ProviderMachineSerializer
+# from .details import QuotaSerializer
+# from .details import SizeSerializer
+# from .details import TagSerializer
+# from .details import UserSerializer
+# from .details import VolumeSerializer
+
+# from .summaries import AllocationSummarySerializer
+# from .summaries import IdentitySummarySerializer
+# from .summaries import ImageSummarySerializer
+# from .summaries import InstanceSummarySerializer
+# from .summaries import ProviderSummarySerializer
+# from .summaries import ProviderMachineSummarySerializer
+# from .summaries import QuotaSummarySerializer
+# from .summaries import SizeSummarySerializer
+# from .summaries import UserSummarySerializer
+# from .summaries import VolumeSummarySerializer
