+++ api/v2/serializers/details/__init__.py
@@ -1,0 +1,13 @@
+from .allocation import AllocationSerializer
+from .identity import IdentitySerializer
+from .image import ImageSerializer
+from .image_bookmark import ImageBookmarkSerializer
+from .instance import InstanceSerializer
+from .project import ProjectSerializer
+from .provider import ProviderSerializer, ProviderTypeSerializer, PlatformTypeSerializer
+from .provider_machine import ProviderMachineSerializer
+from .quota import QuotaSerializer
+from .size import SizeSerializer
+from .tag import TagSerializer
+from .user import UserSerializer
+from .volume import VolumeSerializer
