+++ api/v2/serializers/details/provider_machine.py
@@ -1,24 +1,14 @@
 from core.models import ProviderMachine
 from rest_framework import serializers
-from .image import ImageSummarySerializer
-from .provider import ProviderSummarySerializer
-from .user import UserSerializer
+from ..summaries import ImageSummarySerializer, ProviderSummarySerializer, UserSummarySerializer
 
 
 class ProviderMachineSerializer(serializers.HyperlinkedModelSerializer):
     image = ImageSummarySerializer(source='application')
     provider = ProviderSummarySerializer()
-    created_by = UserSerializer()
+    created_by = UserSummarySerializer()
 
     class Meta:
         model = ProviderMachine
         view_name = 'api_v2:providermachine-detail'
         fields = ('id', 'url', 'image', 'provider', 'created_by', 'start_date', 'end_date')
-
-
-class ProviderMachineSummarySerializer(serializers.HyperlinkedModelSerializer):
-
-    class Meta:
-        model = ProviderMachine
-        view_name = 'api_v2:providermachine-detail'
-        fields = ('id', 'url', 'start_date', 'end_date')
