+++ api/v2/serializers/fields/__init__.py
@@ -1,0 +1,2 @@
+from .provider_machine import ProviderMachineRelatedField
+from .user import UserRelatedField
