+++ api/v2/serializers/summaries/allocation.py
@@ -1,0 +1,9 @@
+from core.models import Allocation
+from rest_framework import serializers
+
+
+class AllocationSummarySerializer(serializers.HyperlinkedModelSerializer):
+    class Meta:
+        model = Allocation
+        view_name = 'api_v2:allocation-detail'
+        fields = ('id', 'url', 'threshold', 'delta')
