+++ api/v2/serializers/summaries/identity.py
@@ -1,0 +1,9 @@
+from core.models import Identity
+from rest_framework import serializers
+
+
+class IdentitySummarySerializer(serializers.HyperlinkedModelSerializer):
+    class Meta:
+        model = Identity
+        view_name = 'api_v2:identity-detail'
+        fields = ('id',)
