+++ api/v2/serializers/summaries/size.py
@@ -1,0 +1,10 @@
+from core.models import Size
+from rest_framework import serializers
+
+
+class SizeSummarySerializer(serializers.HyperlinkedModelSerializer):
+
+    class Meta:
+        model = Size
+        view_name = 'api_v2:size-detail'
+        fields = ('id', 'url', 'alias', 'name', 'cpu', 'disk', 'mem', 'active', 'start_date', 'end_date')
