+++ api/v2/serializers/summaries/tag.py
@@ -1,0 +1,9 @@
+from core.models import Tag
+from rest_framework import serializers
+
+
+class TagSummarySerializer(serializers.HyperlinkedModelSerializer):
+    class Meta:
+        model = Tag
+        view_name = 'api_v2:tag-detail'
+        fields = ('id', 'url', 'name', 'description')
