+++ atmosphere/settings/__init__.py
@@ -508,3 +508,5 @@ if DEBUG:
         from atmosphere.settings.testing import *
     except ImportError:
         pass
+
+TEST_RUNNER = 'django_nose.NoseTestSuiteRunner'
